Removed bfloat16 support for Network predict.

+++ netloader/networks/base.py
@@ -245,7 +245,7 @@ class BaseNetwork:
         with torch.no_grad(), torch.autocast(
                 enabled=self._half,
                 device_type=self._device.type,
-                dtype=torch.bfloat16):
+                dtype=torch.float32):
             for ids, low_dim, high_dim, *_ in loader:
                 in_data, target = self._data_loader_translation(low_dim, high_dim)
                 data.append(pd.DataFrame([
